+++ lib/Elastica/ResultSet.php
@@ -15,21 +15,18 @@ use Elastica\Exception\InvalidException;
 class ResultSet implements \Iterator, \Countable, \ArrayAccess
 {
     /**
-     * Class for the static create method to use.
-     *
-     * @var string
-     */
-    protected static $_class = 'Elastica\\ResultSet';
-
-    /**
      * Results.
      *
-     * @var array Results
+     * @deprecated Accessing this property in an extended class is deprecated. The method will become private in 4.0. Modify results by implementing BuilderInterface and passing a new Builder to your \Elastica\Search instances.
+     *
+     * @var Result[] Results
      */
     protected $_results = array();
 
     /**
      * Current position.
+     *
+     * @deprecated Accessing this property in an extended class is deprecated. The method will become private in 4.0.
      *
      * @var int Current position
      */
@@ -38,94 +35,33 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
     /**
      * Response.
      *
-     * @var \Elastica\Response Response object
+     * @deprecated Accessing this property in an extended class is deprecated. The method will become private in 4.0. Access the response by calling getResponse
+     *
+     * @var Response Response object
      */
-    protected $_response = null;
+    protected $_response;
 
     /**
      * Query.
      *
-     * @var \Elastica\Query Query object
+     * @deprecated Accessing this property in an extended class is deprecated. The method will become private in 4.0. Access the response by calling getQuery
+     *
+     * @var Query Query object
      */
     protected $_query;
 
     /**
-     * @var int
-     */
-    protected $_took = 0;
-
-    /**
-     * @var bool
-     */
-    protected $_timedOut = false;
-
-    /**
-     * @var int
-     */
-    protected $_totalHits = 0;
-
-    /**
-     * @var float
-     */
-    protected $_maxScore = 0;
-
-    /**
      * Constructs ResultSet object.
      *
-     * @param \Elastica\Response $response Response object
-     * @param \Elastica\Query    $query    Query object
+     * @param Response $response Response object
+     * @param Query $query Query object
+     * @param Result[] $results
      */
-    public function __construct(Response $response, Query $query)
+    public function __construct(Response $response, Query $query, $results)
     {
-        $this->rewind();
-        $this->_init($response);
         $this->_query = $query;
-    }
-
-    /**
-     * Creates a new ResultSet object. Can be configured to return a different
-     * implementation of the ResultSet class.
-     *
-     * @param Response $response
-     * @param Query    $query
-     *
-     * @return ResultSet
-     */
-    public static function create(Response $response, Query $query)
-    {
-        $class = static::$_class;
-
-        return new $class($response, $query);
-    }
-
-    /**
-     * Sets the class to be used for the static create method.
-     *
-     * @param string $class
-     */
-    public static function setClass($class)
-    {
-        static::$_class = $class;
-    }
-
-    /**
-     * Loads all data into the results object (initialisation).
-     *
-     * @param \Elastica\Response $response Response object
-     */
-    protected function _init(Response $response)
-    {
         $this->_response = $response;
-        $result = $response->getData();
-        $this->_totalHits = isset($result['hits']['total']) ? $result['hits']['total'] : 0;
-        $this->_maxScore = isset($result['hits']['max_score']) ? $result['hits']['max_score'] : 0;
-        $this->_took = isset($result['took']) ? $result['took'] : 0;
-        $this->_timedOut = !empty($result['timed_out']);
-        if (isset($result['hits']['hits'])) {
-            foreach ($result['hits']['hits'] as $hit) {
-                $this->_results[] = new Result($hit);
-            }
-        }
+        $this->_results = $results;
     }
 
     /**
@@ -227,7 +163,9 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
      */
     public function getTotalHits()
     {
-        return (int) $this->_totalHits;
+        $data = $this->_response->getData();
+
+        return isset($data['hits']['total']) ? (int) $data['hits']['total'] : 0;
     }
 
     /**
@@ -237,7 +175,9 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
      */
     public function getMaxScore()
     {
-        return (float) $this->_maxScore;
+        $data = $this->_response->getData();
+
+        return isset($data['hits']['max_score']) ? (float) $data['hits']['max_score'] : 0;
     }
 
     /**
@@ -247,23 +187,27 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
      */
     public function getTotalTime()
     {
-        return (int) $this->_took;
+        $data = $this->_response->getData();
+
+        return isset($data['took']) ? $data['took'] : 0;
     }
 
     /**
-     * Returns true iff the query has timed out.
+     * Returns true if the query has timed out.
      *
      * @return bool Timed out
      */
     public function hasTimedOut()
     {
-        return (bool) $this->_timedOut;
+        $data = $this->_response->getData();
+
+        return !empty($data['timed_out']);
     }
 
     /**
      * Returns response object.
      *
-     * @return \Elastica\Response Response object
+     * @return Response Response object
      */
     public function getResponse()
     {
@@ -271,7 +215,7 @@ class ResultSet implements \Iterator, \Countable, \ArrayAccess
     }
 
     /**
-     * @return \Elastica\Query
+     * @return Query
      */
     public function getQuery()
     {